requirements.txt
numpy>=1.21.0
flask==2.3.3
flask-cors==4.0.0
flask-compress==1.14
requests==2.31.0
python-dotenv==1.0.0
scikit-learn>=1.0.0
pyvi>=0.1.1
redis>=4.5.0 